pyproject.toml
[project]
name = "algorithmic-trading"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "altair>=5.5.0",
    "awscli>=1.37.14",
    "backtesting>=0.6.1",
    "backtrader[plotting]>=1.9.78.123",
    "boto3>=1.36.14",
    "cython>=3.0.11",
    "eurostat>=1.1.1",
    "fastapi>=0.115.6",
    "fastparquet>=2024.11.0",
    "fastquant>=0.1.8.1",
    "feature-engine>=1.8.3",
    "finplot>=1.9.6",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "joblib>=1.4.2",
    "matplotlib>=3.10.0",
    "motor>=3.5.1",
    "nbformat>=5.10.4",
    "networkx>=3.4.2",
    "pandas>=2.2.3",
    "pandas-datareader>=0.10.0",
    "pandas-market-calendars>=4.6.0",
    "pandas-ta>=0.3.14b0",
    "plotly>=5.24.1",
    "pmdarima>=2.0.4",
    "polars>=1.19.0",
    "pyarrow>=19.0.0",
    "pymongo[srv]>=4.11",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "seaborn>=0.13.2",
    "shap>=0.46.0",
    "skforecast @ git+https://github.com/miguellopezvirues/skforecast.git@master",
    "statsmodels>=0.14.4",
    "ta-lib>=0.6.1",
    "tqdm>=4.67.1",
    "tqdm-joblib>=0.0.4",
    "uvicorn>=0.34.0",
    "yfinance>=0.2.51",
]

[dependency-groups]
dev = [

]
